+++ src/main/kotlin/com/x7ff/rl/replay/api/model/replay/rattletrap/ReplicationActorId.kt
@@ -4,5 +4,5 @@ import com.fasterxml.jackson.annotation.JsonProperty
 
 data class ReplicationActorId (
     @JsonProperty("limit") val limit: Int,
-    @JsonProperty("value") val id: Int
+    @JsonProperty("value") val id: Long
 )
